Got all particles flying
The issue was that GLSL has a byte alignment of 16 bytes for everything.
A bool is 4 bytes, but has 12 bytes of padding.  A vec4 has 16 bytes (4
for each float).  A mat4 has 4 sets of 16 bytes (4x vec4s).  I had to
adjust the structure on the CPU side in order to get the data in the
buffer to line up.